fix: fix npm scripts failing if a parent directory name contains a space
Using $(npm bin) is problematic because it doesn’t escape spaces, which on some operating systems, like macOS, is valid in file and directory names. This causes the “build” and “test” commands to fail if any of the parent directories of “objection-slugify” contain a space.

npm adds the bin path to the PATH environment variable automatically when evaluating scripts, so prepending the path in this context isn’t necessary. (The locally installed versions of these command-line tools will be used instead of any globally installed ones.)